# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  todayPosts: number;
  scheduledToday: number;
  pendingApprovals: number;
  failedPosts: number;
  connectedPlatforms: number;
  totalPlatforms: number;
  queueSize: number;
  draftCount: number;
}

export interface ActivityEvent {
  id: string;
  type: string;
  title: string;
  description?: string;
  timestamp: string;
  user?: string;
  severity?: string;
}

export interface ActivityFeed {
  items: ActivityEvent[];
}

export interface ContentItem {
  id: string;
  title: string;
  body: string;
  type: string;
  specialty: string;
  category: string;
  platform: string[];
  status: string;
  /** @nullable */
  campaign?: string | null;
  /** @nullable */
  brandVoice?: string | null;
  tags: string[];
  /** @nullable */
  scheduledFor?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ContentList {
  items: ContentItem[];
}

export interface ContentInput {
  title: string;
  body: string;
  type: string;
  specialty: string;
  category: string;
  platform: string[];
  status: string;
  /** @nullable */
  campaign?: string | null;
  /** @nullable */
  brandVoice?: string | null;
  tags?: string[];
  /** @nullable */
  scheduledFor?: string | null;
}

export interface ContentUpdate {
  title?: string;
  body?: string;
  type?: string;
  specialty?: string;
  category?: string;
  platform?: string[];
  status?: string;
  /** @nullable */
  campaign?: string | null;
  /** @nullable */
  brandVoice?: string | null;
  tags?: string[];
  /** @nullable */
  scheduledFor?: string | null;
}

export interface QueueItem {
  id: string;
  contentId: string;
  title: string;
  platform: string;
  scheduledFor: string;
  status: string;
  priority: number;
  order: number;
  /** @nullable */
  campaign?: string | null;
  approvalStatus: string;
}

export interface PublishingQueue {
  items: QueueItem[];
  paused: boolean;
}

export interface QueueReorderInput {
  order: number;
}

export interface ScheduledPost {
  id: string;
  contentId: string;
  title: string;
  platform: string;
  scheduledFor: string;
  timezone: string;
  status: string;
  recurring: boolean;
  /** @nullable */
  recurrenceRule?: string | null;
  /** @nullable */
  campaign?: string | null;
}

export interface ScheduledPostList {
  items: ScheduledPost[];
}

export interface ScheduledPostUpdate {
  scheduledFor?: string;
  timezone?: string;
  recurring?: boolean;
  /** @nullable */
  recurrenceRule?: string | null;
}

export interface PlatformConnection {
  id: string;
  platform: string;
  connected: boolean;
  /** @nullable */
  profile?: string | null;
  permissions: string[];
  /** @nullable */
  lastSync?: string | null;
  capabilities: string[];
  /** @nullable */
  tokenExpiresAt?: string | null;
  status: string;
}

export interface PlatformGroup {
  id: string;
  name: string;
  platforms: string[];
}

export interface PlatformList {
  items: PlatformConnection[];
  groups: PlatformGroup[];
}

export interface PlatformConnectionInput {
  platform: string;
  profile?: string;
}

export interface Campaign {
  id: string;
  name: string;
  type: string;
  status: string;
  goal: string;
  progress: number;
  startDate: string;
  endDate: string;
  platforms: string[];
}

export interface CampaignList {
  items: Campaign[];
}

export interface ApprovalComment {
  id: string;
  author: string;
  text: string;
  timestamp: string;
}

export interface ApprovalRecord {
  id: string;
  contentId: string;
  title: string;
  status: string;
  reviewer: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  deadline?: string | null;
  timestamp: string;
  comments?: ApprovalComment[];
}

export interface ApprovalList {
  items: ApprovalRecord[];
}

export interface ApprovalActionInput {
  note: string;
}

export interface PublishingHistoryEntry {
  id: string;
  contentId: string;
  title: string;
  platform: string;
  status: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  campaign?: string | null;
  /** @nullable */
  error?: string | null;
  user: string;
}

export interface PublishingHistoryList {
  items: PublishingHistoryEntry[];
}

/**
 * @nullable
 */
export type NotificationAction = {
  label: string;
  href: string;
} | null;

export interface Notification {
  id: string;
  title: string;
  message: string;
  severity: string;
  category: string;
  timestamp: string;
  read: boolean;
  /** @nullable */
  action?: NotificationAction;
}

export interface NotificationList {
  items: Notification[];
  unreadCount: number;
}

export interface NotificationReadInput {
  ids: string[];
}

export type GetRecentActivityParams = {
limit?: number;
};

export type ListContentParams = {
status?: string;
platform?: string;
campaign?: string;
search?: string;
};

export type GetScheduledPostsParams = {
start?: string;
end?: string;
};

export type ListApprovalsParams = {
status?: string;
};

export type GetPublishingHistoryParams = {
limit?: number;
};

export type ListNotificationsParams = {
unreadOnly?: boolean;
};