# lib/api-zod/src/generated/types/approvalRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalComment } from './approvalComment';

export interface ApprovalRecord {
  id: string;
  contentId: string;
  title: string;
  status: string;
  reviewer: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  deadline?: string | null;
  timestamp: string;
  comments?: ApprovalComment[];
}