# lib/api-zod/src/generated/types/platformGroup.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformGroup {
  id: string;
  name: string;
  platforms: string[];
}